\name{addtest}
\alias{addtest}

\title{
 Add a test set to ada
}

\description{
  \code{addtest} updates the \code{ada} object to have additional testing errors and testing kappa accuracies
  for each iteration.
}

\usage{
  addtest(x,test.x,test.y, \dots)
}

\arguments{
  \item{x}{object generated by the function \code{ada}. }
  \item{test.x}{new x data}
  \item{test.y}{the true labeling for this testing data}
  \item{\dots}{other arguments not used by this function.}
}

\value{
   updated \code{ada} object.
}

\seealso{
  \code{\link{ada}},\code{\link{update.ada}}
}
\keyword{methods}
