% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createVectorInTimeInvX}
\alias{createVectorInTimeInvX}
\title{createVectorInTimeInvX}
\usage{
createVectorInTimeInvX(X)
}
\arguments{
\item{X}{A row from Xstart in timeInvX}
}
\description{
createVectorInTimeInvX
}
\details{
Internal function used in timeInvX
}

