% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_decomposition.R
\name{daily_decomposition}
\alias{daily_decomposition}
\title{Categorizes the time stamps in a data frame daily}
\usage{
daily_decomposition(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Returns the data frame with an additional column referring to the
day time of each time stamp
}
\description{
Categorizes the time stamps in a data frame daily
}
