% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ElectionProb}
\alias{ElectionProb}
\title{ProbNorm}
\usage{
ElectionProb(x, lowLimit = NULL, normalize = TRUE, day = NULL)
}
\arguments{
\item{x}{An "ElectPred" object.}

\item{lowLimit}{Lower limit, parties below are set to 0. If NULL lower limit is 0.}

\item{normalize}{Should the results be normalized (recalculate based on removed parties). If not, "other" is added.}
}
\value{
Returns a dataframe with samples
}
\description{
ProbNorm
}
\details{
Normalizing the probability at electionDay (other parties always counts as 0).
}

