% Generated by roxygen2 (4.0.1): do not edit by hand
\name{electPredict}
\alias{electPredict}
\title{electPredict}
\usage{
electPredict(priors, parameters, data)
}
\arguments{
\item{priors}{A list of prior specifications for the model.}

\item{parameters}{A list of parameter specifications for the model.}

\item{data}{A data.frame with prespecified variables in parameters.}
}
\value{
An electionPredict object that contains both input and estimates.
The stored inputs are:
 \item{\code{data}}{The data.frame used for prediction}
 \item{\code{parameters}}{The parameters used}
 \item{\code{priors}}{The priors used}


The parameters are contains (as samples) in:
 \item{\code{bt}}{The probability distribution for each party and time period (as array)}
 \item{\code{Wb}}{The covariance matrix distribution}
 \item{\code{tau2}}{The house effects distribution}
}
\description{
Creates a multivariate prediction for a multiparty election.
}

