% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Generates evaluation statistic for data frame}
\usage{
evaluate(data_frame, formula, path, write = TRUE)
}
\arguments{
\item{data_frame}{a data frame in specific form [ADA::formatting()]}

\item{formula}{a formula describing the specific layout of the data frame}

\item{path}{a path with file name for the generated .csv file. If no path but
name is provided the .csv file will be generated in the current working
directory}

\item{write}{determines if .csv file should be generated}
}
\value{
returns summary statistic in convenient form
}
\description{
creates a statistic containing \code{mean}, \code{sd} etc. of the
data frame provided. The data frame must be formatted with [ADA::formatting()]
beforehand. Depending on the number of levels in season and time of day the size
of the resulting statistic can vary.
}
\seealso{
[ADA::formatting()], [base::data.frame()]
}
