% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expandPollToPeriod}
\alias{expandPollToPeriod}
\title{expandPollToPeriod}
\usage{
expandPollToPeriod(polls = Y_all)
}
\arguments{
\item{polls}{Dataframe with (at least) variables periodFrom, periodTo, house, time, n}
}
\value{
Expanded dataset with one poll per day
}
\description{
expandPollToPeriod
}
\details{
Expand a poll to one observation each day. Missing houses are removed.
}

