% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExpectCorr}
\alias{ExpectCorr}
\title{ExpectCorr}
\usage{
ExpectCorr(x)
}
\arguments{
\item{x}{An "ElectPred" object.}
}
\value{
Returns a vector with the expected prediction.
}
\description{
ExpectCorr
}
\details{
Expected state Covariance Matrix
}

