% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExpectCovariance}
\alias{ExpectCovariance}
\title{ExpectCovariance}
\usage{
ExpectCovariance(x)
}
\arguments{
\item{x}{An "ElectPred" object.}
}
\value{
Returns a vector with the expected prediction.
}
\description{
ExpectCovariance
}
\details{
Expected state Covariance Matrix
}

