% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExpectPred}
\alias{ExpectPred}
\title{ExpectPred}
\usage{
ExpectPred(x, day = NULL)
}
\arguments{
\item{obj}{An "ElectPred" object.}
}
\value{
Returns a vector with the expected prediction.
}
\description{
ExpectPred
}
\details{
Analyze function to calculate the expected prediction at the election.
}

