% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExpectStDev}
\alias{ExpectStDev}
\title{ExpectStDev}
\usage{
ExpectStDev(x)
}
\arguments{
\item{x}{An "ElectPred" object.}
}
\value{
Returns a vector with the expected prediction.
}
\description{
ExpectStDev
}
\details{
Expected state Covariance Matrix
}

