% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting}
\alias{formatting}
\title{Formats a data frame in specific format}
\usage{
formatting(data_frame, formula, tz = "UTC", format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{data_frame}{a data frame}

\item{formula}{a formula describing the specific layout of the data frame}

\item{tz}{a time zone like "UTC", "MET", ...}

\item{format}{the format in which the time stamps are given in the data frame}
}
\value{
returns a data frame which has a special format
}
\description{
generates a new data frame according to the formula provided and
adds columns referring to the specific season and time of day (tod) for each
time stamp. Pollutants will be on the left side, the date terms on the right
including season and tod. The explanatory variables will be in the middle.
}
\seealso{
[base::data.frame()], [stats::formula()]
}
