% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.ElectPred}
\alias{plot.ElectPred}
\title{plot.ElectPred}
\usage{
\method{plot}{ElectPred}(x, party = "all", showPriors = TRUE,
  intervals = c(0.95, 0.75), returnPlots = FALSE, ...)
}
\arguments{
\item{x}{An ElectPredict object}

\item{party}{Which party to plot}

\item{showPriors}{Should the prior on electionday be shown?}

\item{intervals}{The probability intervals to be calculated}

\item{returnPlots}{Don't print the plot, return them as list withh ggplot objects}

\item{...}{not currently in use.}
}
\value{
If returnPlots=TRUE returneras en lista med ggplotobjekten.
}
\description{
plot.ElectPred
}
\details{
Plots how the states and polls over time for each party
}

