% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rinvgamma}
\alias{rinvgamma}
\title{rinvgamma}
\usage{
rinvgamma(n, shape, scale = 1)
}
\arguments{
\item{n}{See n in rgamma()}

\item{shape}{Alpha}

\item{scale}{Beta}
}
\description{
rinvgamma
}
\details{
Samples an inverse gamma (taken from MCMCpack)
}

