% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleMissingYnorm}
\alias{sampleMissingYnorm}
\title{sampleMissingYnorm}
\usage{
sampleMissingYnorm(statesMissing, samplesize = 500)
}
\arguments{
\item{statesMissing}{Matrix with missing states to sample from}

\item{samplesize}{The assumed sample size to sample from as a normal distribution}
}
\value{
A matrix where values for Y has been sampled.
}
\description{
sampleMissingYnorm
}
\details{
Samples missing values of Y. Assuming a MVN distribution (with Covar as multinomial)
}

