% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleTau2}
\alias{sampleTau2}
\title{sampleTau2}
\usage{
sampleTau2(prior, Yt, bt)
}
\arguments{
\item{prior}{A 2 * h matrix containing alpha and beta as columns and a prior for each house}

\item{Yt}{Observation matrix (rows are observations and columns are parties)}

\item{bt}{State matrix (rows are observations and columns are parties)}
}
\value{
A matrix with each matrix as a row
}
\description{
sampleTau2
}
\details{
Samples tau^2, the house design effect in the model
}

