% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_decomposition.R
\name{seasonal_decomposition}
\alias{seasonal_decomposition}
\title{Categorizes the time stamps in a data frame seasonally}
\usage{
seasonal_decomposition(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Returns the data frame with an additional column referring to the
season of each time stamp
}
\description{
Categorizes the time stamps in a data frame seasonally
}
\examples{
y = as.Date("2019-12-17") + 1:5
y = data
y = seasonal_decomposition(y)
}
