% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.ElectPred}
\alias{summary.ElectPred}
\title{summary}
\usage{
\method{summary}{ElectPred}(x, lowLimit = 0.04, day = NULL, alpha = 0.05)
}
\arguments{
\item{x}{An "ElectPred" object.}
}
\value{
Returns a vector with the expected prediction.
}
\description{
summary
}
\details{
Summation of interest results.
}

