% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timeInvX}
\alias{timeInvX}
\title{timeInvX}
\usage{
timeInvX(pMatrix, n)
}
\arguments{
\item{pMatrix}{Matrix with proportions as rows}

\item{n}{vector vith no of observations}
}
\value{
A matrix with each matrix as a row
}
\description{
timeInvX
}
\details{
Calculates the time variant V matrix
}

