\name{soldat}
\alias{soldat}
\docType{data}
\title{ Solubility Data }
\description{
	A data set that contains information about compounds used in drug discovery. 
      Specifically, this data set consists of 5631 compounds on which an in-house
      solubility screen (ability of a compound to dissolve in a water/solvent mixture) was performed. 

     Based on this screen, compounds were categorized as either insoluble (n=3493) or soluble (n=2138).  
     Then, for each compound, 72 continuous, noisy structural
     descriptors were computed.  
}
\usage{data(soldat)}
\format{
  A data frame with 5631 observations on the following 73 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{x5}}{a numeric vector}
    \item{\code{x6}}{a numeric vector}
    \item{\code{x7}}{a numeric vector}
    \item{\code{x8}}{a numeric vector}
    \item{\code{x9}}{a numeric vector}
    \item{\code{x10}}{a numeric vector}
    \item{\code{x11}}{a numeric vector}
    \item{\code{x12}}{a numeric vector}
    \item{\code{x13}}{a numeric vector}
    \item{\code{x14}}{a numeric vector}
    \item{\code{x15}}{a numeric vector}
    \item{\code{x16}}{a numeric vector}
    \item{\code{x17}}{a numeric vector}
    \item{\code{x18}}{a numeric vector}
    \item{\code{x19}}{a numeric vector}
    \item{\code{x20}}{a numeric vector}
    \item{\code{x21}}{a numeric vector}
    \item{\code{x22}}{a numeric vector}
    \item{\code{x23}}{a numeric vector}
    \item{\code{x24}}{a numeric vector}
    \item{\code{x25}}{a numeric vector}
    \item{\code{x26}}{a numeric vector}
    \item{\code{x27}}{a numeric vector}
    \item{\code{x28}}{a numeric vector}
    \item{\code{x29}}{a numeric vector}
    \item{\code{x30}}{a numeric vector}
    \item{\code{x31}}{a numeric vector}
    \item{\code{x32}}{a numeric vector}
    \item{\code{x33}}{a numeric vector}
    \item{\code{x34}}{a numeric vector}
    \item{\code{x35}}{a numeric vector}
    \item{\code{x36}}{a numeric vector}
    \item{\code{x37}}{a numeric vector}
    \item{\code{x38}}{a numeric vector}
    \item{\code{x39}}{a numeric vector}
    \item{\code{x40}}{a numeric vector}
    \item{\code{x41}}{a numeric vector}
    \item{\code{x42}}{a numeric vector}
    \item{\code{x43}}{a numeric vector}
    \item{\code{x44}}{a numeric vector}
    \item{\code{x45}}{a numeric vector}
    \item{\code{x46}}{a numeric vector}
    \item{\code{x47}}{a numeric vector}
    \item{\code{x48}}{a numeric vector}
    \item{\code{x49}}{a numeric vector}
    \item{\code{x50}}{a numeric vector}
    \item{\code{x51}}{a numeric vector}
    \item{\code{x52}}{a numeric vector}
    \item{\code{x53}}{a numeric vector}
    \item{\code{x54}}{a numeric vector}
    \item{\code{x55}}{a numeric vector}
    \item{\code{x56}}{a numeric vector}
    \item{\code{x57}}{a numeric vector}
    \item{\code{x58}}{a numeric vector}
    \item{\code{x59}}{a numeric vector}
    \item{\code{x60}}{a numeric vector}
    \item{\code{x61}}{a numeric vector}
    \item{\code{x62}}{a numeric vector}
    \item{\code{x63}}{a numeric vector}
    \item{\code{x64}}{a numeric vector}
    \item{\code{x65}}{a numeric vector}
    \item{\code{x66}}{a numeric vector}
    \item{\code{x67}}{a numeric vector}
    \item{\code{x68}}{a numeric vector}
    \item{\code{x69}}{a numeric vector}
    \item{\code{x70}}{a numeric vector}
    \item{\code{x71}}{a numeric vector}
    \item{\code{x72}}{a numeric vector with missing data}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
     data(soldat)
}
\keyword{datasets}
